Fix a bunch of compiler warnings.

+++ IO/src/ImageLoaderPNG.h
@@ -11,7 +11,7 @@ class ImageLoaderPNG : public ImageLoader {
   virtual bool ReadData();
 
 private:
-  unsigned int m_StreamPosition;
+  uint64 m_StreamPosition;
   friend class PNGStreamReader;
 };
