+++ IO/src/ImageWriterPNG.h
@@ -12,7 +12,7 @@ class ImageWriterPNG : public ImageWriter {
 
   virtual bool WriteImage();
  private:
-	uint32 m_StreamPosition;
+	uint64 m_StreamPosition;
 	uint32 m_TotalBytesWritten;
 	friend class PNGStreamWriter;
 };
